requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2===2.9.9
Werkzeug==2.3.7
PyJWT==2.8.0
python-dotenv==1.0.0
password-strength==0.0.3.post2
validate_docbr==1.10.0
Pillow==10.0.1
psycopg2-binary===2.9.9
stripe===7.4.0
Flask-Mail===0.9.1
